moved IDs to Long for Lab 8

+++ Lab-08-solution/src/main/java/com/petclinic/core/OwnerService.java
@@ -26,7 +26,7 @@ public class OwnerService {
         else return owner.get();
     }
 
-    public Owner findById(Integer id) {
+    public Owner findById(Long id) {
         Optional<Owner> owner = ownerRepository.findById(id);
         if (owner.isEmpty()) {
             throw new EntityNotFoundException("could not find Owner with id: " + id);
